klientctl: move metrics tag generator to a separate package
