+++ go/src/koding/klientctl/commands/cli/cobra.go
@@ -10,7 +10,7 @@ type CobraFuncE func(cmd *cobra.Command, args []string) error
 // CobraCmdMiddleware defines function that modifies cobra commands in order
 // to add additional functionality to their handlers. All middlewares in this
 // package must use only cobra RunE function to not break other components.
-type CobraCmdMiddleware func(cli *CLI, cmd *cobra.Command)
+type CobraCmdMiddleware func(cli *CLI, rootCmd *cobra.Command)
 
 // MultiCobraCmdMiddleware creates a middleware that applies all provided
 // functions preserving invocation order.
